Pull upstream

+++ src/c72_menuHA_GPS.hpp
@@ -210,7 +210,7 @@ void printHASubmenu()
     }
     else if (haState == STARTING_GPS)
     {
-        sprintf(satBuffer, "  Found %d sats", (unsigned int)gps.satellites.value());
+        sprintf(satBuffer, "  Found %lu sats", gps.satellites.value());
         satBuffer[0] = ind[indicator];
     }
     else if (haState == ENTER_HA_MANUALLY)
